moved database out
